Rename : 디렉토리 정리 및 컴포넌트 import문 수정
