+++ src/Basic/Avatar/Avatar.stories.js
@@ -1,5 +1,4 @@
 import React from "react";
-
 import Avatar from "./";
 
 export default {
