+++ src/Basic/Typography/Typography.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
+
 import { color, typography } from "../../shared/styles";
 
 const textHeight = {
