+++ src/Basic/Typography/index.js
@@ -1,1 +1,1 @@
-export { default as Typography } from "./Typography";
+export { default } from "./Typography";
