+++ src/Components/index.js
@@ -1,10 +1,10 @@
-// export { default as Badge } from "./Badge";
-// export { default as Button } from "./Button";
+export { default as Badge } from "./Badge";
+export { default as Button } from "./Button";
 export { default as Comment } from "./Comment";
-// export { default as ImageBox } from "./ImageBox";
+export { default as ImageBox } from "./ImageBox";
 export { default as Input } from "./Input";
 export { default as NavItem } from "./NavItem";
-// export { default as SideBarItem } from "./SideBarItem";
-// export { default as SubNavbarItem } from "./SubNavbarItem";
+export { default as SideBarItem } from "./SideBarItem";
+export { default as SubNavbarItem } from "./SubNavbarItem";
 export { default as TocItem } from "./TocItem";
 export { default as Toggle } from "./TocItem";
