+++ src/Layouts/index.js
@@ -1,6 +1,6 @@
-// export { default as Card } from "./Card";
+export { default as Card } from "./Card";
 export { default as Navbar } from "./Navbar";
-// export { default as Sidebar } from "./Sidebar";
-// export { default as SubNavbar } from "./SubNavbar";
+export { default as Sidebar } from "./Sidebar";
+export { default as SubNavbar } from "./SubNavbar";
 export { default as Toc } from "./Toc";
 export { default as TocNav } from "./TocNav";
